+++ lib/ecto_shorts/common_filters/builder.ex
@@ -158,10 +158,6 @@ defmodule EctoShorts.CommonFilters.Builder do
     Lock.build_query(:lock, source, query, selected_binding, term, opts)
   end
 
-  defp apply_filter(:first, source, query, selected_binding, term, opts) do
-    Limit.build_query(:limit, source, query, selected_binding, term, opts)
-  end
-
   defp apply_filter(:limit, source, query, selected_binding, term, opts) do
     Limit.build_query(:limit, source, query, selected_binding, term, opts)
   end